pyproject.toml
[project]
name = "kakehashi"
version = "0.1.0"
description = "X (Twitter) to Misskey crosspost bot"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "twikit @ git+https://github.com/cmj/twikit.git",
    "httpx>=0.28.1",
    "pillow>=11.2.1",
    "pillow-avif-plugin>=1.4.6",
]

[project.optional-dependencies]
# config.toml の url_cleaner = true 時に必要
url-cleaner = [
    "url-cleaner>=0.1.5",
]

[tool.uv]
python-preference = "only-managed"
